==> 07_hts_api.ipynb <==
# 7. 증권사 API 연결 및 매매


증권사 API?

비교

### 7.1 아나콘다에서 32비트 가상환경 만들기

먼저 키움 API를 사용하기 위해서는 64비트가 아닌 32비트 환경이 필요합니다. 따라서 Anaconda prompt 실행하시고 다음과 같이 입력합니다.

```
conda info
```

위 코드의 결과를 보면 현재 64비트임을 알 수 있습니다. 따라서 아래와 같은 코드를 입력해 32비트로 변경합니다.

```
set CONDA_FORCE_32BIT=1
conda info
```

만약 다시 64비트로 변경하고 싶다면 다음과 같이 입력합니다.

```
set CONDA_FORCE_32BIT=0
```

아나콘다를 32비트로 바꾸었으면 32비트 가상환경을 만듭니다. 현재 사용중인 파이썬 버전이 3.8.5 이므로, 32비트 가상환경을 만들기 위해 다음과 같이 입력합니다. python= 뒤에 본인이 사용중인 파이썬 버전을 입력하시면 됩니다.

```
conda create --name py38_32 python=3.8.5 
conda activate py38_32 
```

## 7.2 파이참에서 32비트 환경 설정하기

https://stynxh.github.io/2020-09-23-python-path-for-anaconda-venv-in-win10-korean/

where python 통해 가상환경 확인하기

Conda Environment를 클릭하고 앞서만든 가상환경을 연동시켜줍니다. 

그리고 실행된 파이참 파일에서 파이썬이 32비트인지 확인합니다.

```
import platform
print(platform.architecture())
```

## 7.3 키움 계좌개설 및 모듈 설치

Blah Blah

## 7.4 PyQT 설치하기
